+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+#### 1.1.0 (1/23/18)
 * Added: Set default grid target to "_blank" and rel to "noopener".
 * Changed: Only run updater in the admin.